refactor: move endpoint to appropriate class
